Create TEST 2
Question 1) A java class is a program that is used to contain the set of methods that performs the class’s tasks. The methods would perform tasks such as generate a random number, or receive user input.

Question 2) Declaring a variable is simply creating an object that has a "type" and a "name". The type can range from simple data types such as "int" to complex classes. The name is simply used to tell the compiler that you will be accessing that data type using that name.

Question 3-7) IN CODE BOX